+++ YL1CC3_HFT_2022231.Logic/CarLogic.cs
@@ -30,17 +30,35 @@ namespace YL1CC3_HFT_2022231.Logic
 
         public void Create(Car item)
         {
+            if (item.Price<0)
+            {
+                throw new ArgumentException();
+            }
             this.repo.Create(item);
         }
 
         public void Delete(int id)
         {
-            this.repo.Delete(id);
+            if (id < 1)
+            {
+                throw new ArgumentOutOfRangeException();
+            }
+            else
+            {
+                this.repo.Delete(id);
+            }
         }
 
         public Car Read(int id)
         {
-            return this.repo.Read(id);
+            if (id<1)
+            {
+                throw new ArgumentOutOfRangeException();
+            }
+            else
+            {
+                return this.repo.Read(id);
+            }
         }
 
         public IQueryable<Car> ReadAll()
@@ -50,6 +68,10 @@ namespace YL1CC3_HFT_2022231.Logic
 
         public void Update(Car item)
         {
+            if (item.Price < 0)
+            {
+                throw new ArgumentException();
+            }
             this.repo.Update(item);
         }
