+++ YL1CC3_HFT_2022231.Logic/RentLogic.cs
@@ -31,17 +31,35 @@ namespace YL1CC3_HFT_2022231.Logic
 
         public void Create(Rent item)
         {
+            if (item.Interval<0)
+            {
+                throw new ArgumentException();
+            }
             this.repo.Create(item);
         }
 
         public void Delete(int id)
         {
-            this.repo.Delete(id);
+            if (id < 1)
+            {
+                throw new ArgumentOutOfRangeException();
+            }
+            else
+            {
+                this.repo.Delete(id);
+            }
         }
 
         public Rent Read(int id)
         {
-            return this.repo.Read(id);
+            if (id < 1)
+            {
+                throw new ArgumentOutOfRangeException();
+            }
+            else
+            {
+                return this.repo.Read(id);
+            }
         }
 
         public IQueryable<Rent> ReadAll()
@@ -51,6 +69,10 @@ namespace YL1CC3_HFT_2022231.Logic
 
         public void Update(Rent item)
         {
+            if (item.Interval < 0)
+            {
+                throw new ArgumentException();
+            }
             this.repo.Update(item);
         }
     }
